refactor all function based views to ViewSets

+++ django/project/views.py
@@ -30,7 +30,8 @@ from .project_data import project_structure
 
 class ProjectPublicViewSet(ViewSet):
 
-    def by_district(self, request, country_id):
+    @staticmethod
+    def by_district(request, country_id):
         """
         Retrieves list of projects by district
         """
@@ -67,7 +68,8 @@ class ProjectPublicViewSet(ViewSet):
 
         return Response(result_dict)
 
-    def list_all(self, request, *args, **kwargs):
+    @staticmethod
+    def list_all(request, *args, **kwargs):
         """
         Retrieves list of projects (optionally by country)
         """
@@ -95,6 +97,12 @@ class ProjectPublicViewSet(ViewSet):
         }], projects, result_list)
 
         return Response(result_list)
+
+    @staticmethod
+    def project_structure(request):
+        countries = [dict(id=x.id, name=x.name) for x in Country.objects.all()]
+        project_structure.update(countries=countries)
+        return Response(project_structure)
 
 
 class ProjectListViewSet(TokenAuthMixin, ViewSet):
@@ -203,6 +211,63 @@ class ProjectGroupViewSet(TeamTokenAuthMixin, RetrieveModelMixin, GenericViewSet
         return Response(serializer.data)
 
 
+class ProjectVersionViewSet(TeamTokenAuthMixin, ViewSet):
+
+    def create(self, request, project_id):
+        """
+        Makes versions out of Toolkit and coverage data for the project.
+        """
+        project = get_object_or_400(Project, "No such project.", id=project_id)
+        self.check_object_permissions(request, project)
+
+        last_cov_ver = CoverageVersion.objects.filter(project_id=project_id).order_by("-version").first()
+        if not last_cov_ver:
+            # No versions yet.
+            new_version = 1
+        else:
+            new_version = last_cov_ver.version + 1
+        current_cov = project.data["coverage"]
+        new_cov_ver = CoverageVersion(
+                            project_id=project_id,
+                            version=new_version,
+                            data=current_cov)
+        new_cov_ver.save()
+
+        # Make a new version from current toolkit.
+        last_toolkit_ver = ToolkitVersion.objects.filter(project_id=project_id).order_by("-version").first()
+        if not last_toolkit_ver:
+            # No versions yet.
+            new_version = 1
+        else:
+            new_version = last_toolkit_ver.version + 1
+        current_toolkit = get_object_or_400(Toolkit, "No such Toolkit", project_id=project_id).data
+        new_toolkit_ver = ToolkitVersion(
+                            project_id=project_id,
+                            version=new_version,
+                            data=current_toolkit)
+        new_toolkit_ver.save()
+        return Response(status=status.HTTP_201_CREATED)
+
+    def toolkit_versions(self, request, project_id):
+        """
+        Retrieves all toolkit versions for the given project_id.
+        """
+        project = get_object_or_400(Project, "No such project.", id=project_id)
+        self.check_object_permissions(request, project)
+
+        toolkit_versions = ToolkitVersion.objects.filter(project_id=project_id) \
+            .order_by("version").values("version", "data", "modified")
+        return Response(toolkit_versions)
+
+    def coverage_versions(self, request, project_id):
+        """
+        Retrieves all coverage versions for the given project_id.
+        """
+        coverage_versions = CoverageVersion.objects.filter(project_id=project_id) \
+            .order_by("version").values("version", "data", "modified")
+        return Response(coverage_versions)
+
+
 class PartnerLogoListViewSet(ViewSet):
 
     def list(self, request, project_id):
@@ -220,6 +285,8 @@ class PartnerLogoViewSet(TeamTokenAuthMixin, ViewSet):
         Creates partnerlogos from the uploaded files.
         """
         project = get_object_or_400(Project, "No such project.", id=project_id)
+        self.check_object_permissions(self.request, project)
+
         logos = []
         # Get and store binary files for partnerlogos.
         for key, value in request.FILES.items():
@@ -228,7 +295,9 @@ class PartnerLogoViewSet(TeamTokenAuthMixin, ViewSet):
         return Response(logos)
 
     def destroy(self, request, pk=None):
-        get_object_or_400(PartnerLogo, "No such logo.", id=pk).delete()
+        partner_logo = get_object_or_400(PartnerLogo, "No such logo.", id=pk)
+        self.check_object_permissions(self.request, partner_logo.project)
+        partner_logo.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -243,6 +312,8 @@ class FilePostViewSet(TeamTokenAuthMixin, ViewSet):
 
     def create(self, request, project_id):
         project = get_object_or_400(Project, "No such project.", id=project_id)
+        self.check_object_permissions(self.request, project)
+
         # Get and store binary files for publications and reports.
         files = []
         file_type = None
@@ -271,79 +342,6 @@ class FileDeleteViewSet(TeamTokenAuthMixin, ViewSet):
 
     def destroy(self, request, pk):
         file = get_object_or_400(File, "No such file.", id=pk)
+        self.check_object_permissions(self.request, file.project)
         file.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-@api_view(['POST'])
-@authentication_classes((TokenAuthentication,))
-@permission_classes((IsAuthenticated, InTeamOrReadOnly))
-def make_version(request, project_id):
-    """
-    Makes versions out of Toolkit and coverage data for the project.
-
-    Args:
-        project_id: id of the project.
-    """
-    # Make a new version from current coverage.
-    last_cov_ver = CoverageVersion.objects.filter(project_id=project_id).order_by("-version").first()
-    if not last_cov_ver:
-        # No versions yet.
-        new_version = 1
-    else:
-        new_version = last_cov_ver.version + 1
-    current_cov = get_object_or_400(Project, "No such project.", id=project_id).data["coverage"]
-    new_cov_ver = CoverageVersion(
-                        project_id=project_id,
-                        version=new_version,
-                        data=current_cov)
-    new_cov_ver.save()
-
-    # Make a new version from current toolkit.
-    last_toolkit_ver = ToolkitVersion.objects.filter(project_id=project_id).order_by("-version").first()
-    if not last_toolkit_ver:
-        # No versions yet.
-        new_version = 1
-    else:
-        new_version = last_toolkit_ver.version + 1
-    current_toolkit = get_object_or_400(Toolkit, "No such Toolkit", project_id=project_id).data
-    new_toolkit_ver = ToolkitVersion(
-                        project_id=project_id,
-                        version=new_version,
-                        data=current_toolkit)
-    new_toolkit_ver.save()
-    return Response(status=status.HTTP_201_CREATED)
-
-
-@api_view(['GET'])
-@authentication_classes((TokenAuthentication,))
-@permission_classes((IsAuthenticated,))
-def get_coverage_versions(request, project_id):
-    """
-    Retrieves all coverage versions for the given project_id.
-    """
-    coverage_versions = CoverageVersion.objects.filter(project_id=project_id) \
-                            .order_by("version").values("version", "data", "modified")
-    return Response(coverage_versions)
-
-
-@api_view(['GET'])
-@authentication_classes((TokenAuthentication,))
-@permission_classes((IsAuthenticated, InTeamOrReadOnly))
-def get_toolkit_versions(request, project_id):
-    """
-    Retrieves all toolkit versions for the given project_id.
-    """
-    toolkit_versions = ToolkitVersion.objects.filter(project_id=project_id) \
-                            .order_by("version").values("version", "data", "modified")
-    return Response(toolkit_versions)
-
-
-@api_view(['GET'])
-def get_project_structure(request):
-    """
-    View for providing form data for create/edit project.
-    """
-    countries = [dict(id=x.id, name=x.name) for x in Country.objects.all()]
-    project_structure.update(countries=countries)
-    return Response(project_structure)
